Explore sendRequest API with handler function
Gregory Collins suggested that the signature for recieveResponse would
take a handler function as opposed to just an InputStream. I didn't
understand that at the time, but I'm seeing it now: we're handing the
user the stream to write to — then read from — the server, rather than
asking them to hand us a stream they build themselves.

Experiment with this idea, first with sendRequest.

Add a stub for setting the Content-Type of a Request.

Rough out the very beginnings of the code to compose the actual HTTP
message for sending.