Begin implementing ResponseParser
Preliminary steps down the road of writing code to parse HTTP
responses.

Start using 'build-test' path in Makefile; obviously this is
suppsoed to be the proper test suite but one thing at a time. At the
moment we can just test against an example HTTP response in a text
file.