Move Request serialization code to Types
Although it strictly has more to do with the Connection than the
Request, move the code converting the request metadata to HTTP format
into Network.Http.Types. Code there will eventually need to have the
usual performance critical machinations, making it messy, but this has
the advantage that we can (mis)use it to power the Show instance of
Request for debugging.