Fix chunked encoding parser
When handling a chunked transfer encoding, our parser had a serious
bug that it failed to look for the final terminating CRLF pair. Got
away with this on single requests most of the time, but pipelining
requests showed up a parser crash.

Thanks to Gregory Collins for helping debug the issue and for having
fixed a corresponding problem in io-streams.